// src/app/Models/GetAttendCalendarSimplifyData.ts
// Generated by https://quicktype.io

export interface GetAttendCalendarSimplifyData {
    EmpID:             string;
    AttendDate:        string;
    ActualRoteID:      number;
    ActualRoteCode:    string;
    ActualRoteName:    string;
    OnTime:            string;
    OffTime:           string;
    AttendExceptional: boolean;
    Abs:               boolean;
    AbsFlow:           boolean;
    Holiday:           boolean;
}